style(permission): selectable access-mode rows, centered settings footer
- AccessModeSelector rows now behave like the options they are: hover
  tint, brand-tinted background while checked, and the label kept on
  one line so only the description truncates. Disabled rows opt out of
  the hover tint so they still read as unavailable.
- That truncated description gets the same hover-only tooltip the
  member list uses, so TruncatedTooltip moves out of PermissionListTab
  into its own module. Its polymorphic ref keeps one `any` — now with
  an inline reason — and the file's suppression entry shrinks.
- Both settings pages center their footer actions unconditionally; the
  knowledge page had centered only in create mode, so the same footer
  jumped between layouts depending on how you entered it.
